Moreutils: Added ifne
Hi, I read the suggestion about ifne in the Discussion page of
moreutils and found it interesting. It's a command that runs another
process if the standard input is not empty. I think it can be very
useful in admin scripts so I've implemented it. I attach a patch
against the current head of moreutils' git. You may include it in
moreutils if you think it's worth it.

I've considered making an "ifempty" or "ifne -v" that would run a
command if the standard input is empty, but I'm not sure if that could
be useful. Any comments will be appreciated. Regards,

Javi