patch for moreutils Makefile
Hi,

I just packaged up moreutils for Gentoo and added it to their system,
I'll go add it to the SuSE repo system next week as well, thanks a lot
for the utils.

Here's a small patch to the makefile to make it easier for distros to
change the docbook2x-man binary as it is named other things (Gentoo for
example has it as docbook2man.pl).  This way it can be changed on the
build line, and the Makefile doesn't have to be patched.

Thanks for already fixing the DESTDIR thing, that also will make things
easier.

And also, you mention dropping you an email if you want to be notified
when new versions come out, I'd appreciate it if you could add me to
that list if you have one.

thanks again,

greg k-h

--------------------
Subject: Make docbook2x-man able to be overridden in Makefile

This makes it easier for distros that do not have the docbook2x-man
binary with that name, override it on the command line instead of having
to patch the Makefile.