parallel: Implement -l (maxload) and -i (replace {})
Add support for only starting jobs when the load average is low
enough, this is done with -l and acts similarly to how make does it.

In order to construct more complex commands, add -i which will make
parallel replace {} with the argument at an arbitrary place in the
command line.